### 2023-01-29 #5113 @editorialbot
[PRE REVIEW]: SDGdetector: an R-based text mining tool for quantifying efforts toward Sustainable Development Goals
**Submitting author:** @Yingjie4Science (<a href="http://orcid.org/0000-0002-8401-0649">Yingjie Li</a>)
**Repository:** https://github.com/Yingjie4Science/SDGdetector
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
