### 2023-02-01 #5113 @fboehm
@caleblucas @colbrydi @giswqs - would you be willing to review this submission for the Journal of Open Source Software (JOSS)? JOSS reviews are checklist-driven. You can read more about the review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
I'm happy to address any questions that you might have. Thanks!
