### 2023-02-01 #5113 @giswqs
I am not an active R user. I would be more comfortable with reviewing Python packages. Here are some rspatial experts:

- https://github.com/csaybar
- https://github.com/brownag
- https://github.com/rafapereirabr
