### 2023-02-01 #5113 @caleblucas
I’m happy to review. Thanks,CalebOn Jan 31, 2023, at 20:46, Qiusheng Wu ***@***.***> wrote:﻿
I am not an active R user. I would be more comfortable with reviewing Python packages. Here are some rspatial experts:

https://github.com/csaybar
https://github.com/brownag
https://github.com/rafapereirabr
