### 2023-02-03 #5113 @fboehm
@rafapereirabr would you be willing to review this submission for JOSS? JOSS reviews are checklist-driven. You can read more about the review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
I'm happy to address any questions that you might have. Thanks!
