### 2023-02-03 #5124 @editorialbot
[REVIEW]: SDGdetector: an R-based text mining tool for quantifying efforts toward Sustainable Development Goals
**Submitting author:** @Yingjie4Science (<a href="http://orcid.org/0000-0002-8401-0649">Yingjie Li</a>)
**Repository:** https://github.com/Yingjie4Science/SDGdetector
**Branch with paper.md** (empty if default branch):
**Version:** v2.7.2
