### 2023-02-14 3cbb5af
clarify the differences to similar work - text2sdg
