### 2023-02-14 Yingjie4Science/SDGdetector#4 @csaybar
JOSS Review - csaybar
Hi @Yingjie4Science,

I'm having a lot of struggle reviewing your package/paper.  Sorry maybe is my fault (it's my first time as a JOSS reviewer 😬).
On the JOSS website it says **Version: v0.1.0**  (https://github.com/openjournals/joss-reviews/issues/5124) but this version does not have the functionality you describe in the paper. It looks like a very early version of your package.

The latest version doesn't install on my computer.

```r
remotes::install_github("Yingjie4Science/SDGdetector")
Warning in file(filename, "r", encoding = encoding) :
  cannot open file './Code/function_lookaround_nearby_n.R': No such file or directory
Error in file(filename, "r", encoding = encoding) :
  cannot open the connection
ERROR: lazydata failed for package ‘SDGdetector’
```

You should check  this R file: https://github.com/Yingjie4Science/SDGdetector/blob/ed5358ba7456f5504cc106aa973cbfedcad2ab5e/data/helper_SDG_search_terms.R#L92

I would really appreciate  a new release of SDGdetector with the code to review.

Best!
