### 2023-02-17 #5124 @csaybar
Hi @fboehm, the version to review is v0.1.0, but this version differs from the one described in the paper. The latest version on GitHub (2.3.0):

```r
remotes::install_github("Yingjie4Science/SDGdetector")
```

```
Warning in file(filename, "r", encoding = encoding) :
  cannot open file './Code/function_lookaround_nearby_n.R': No such file or directory
Error in file(filename, "r", encoding = encoding) :
  cannot open the connection
ERROR: lazydata failed for package ‘SDGdetector’
* removing ‘/home/csaybar/R/x86_64-pc-linux-gnu-library/4.0/SDGdetector’
Warning message:
  In i.p(...) :
  installation of package ‘/tmp/RtmpU47SxK/file48d12db43451/SDGdetector_2.3.0.tar.gz’ had non-zero exit status
```

It doesn't install because `function_lookaround_nearby_n.R` doesn't exist. I opened an issue here:
https://github.com/Yingjie4Science/SDGdetector/issues/4
