### 2023-02-18 da855bd
fix error due to 'function_lookaround_nearby_n.R'
