+++ data/helper_SDG_search_terms.R
@@ -4,22 +4,21 @@
 ## ######################################################################## ##
 
 #' **Last update on**:  3/31/2022
-#' 
-#' **New changes**: 
-#' 
+#'
+#' **New changes**:
+#'
 #' Compare to the earlier version, we made the following changes
-#' 
-#' 1. Instead of combining multiple term lists by `OR` for one particular target,  
-#'    it is more intuitive and accurate to add each alternative term list to the search 
-#'    term table or database directly. 
-#'    
+#'
+#' 1. Instead of combining multiple term lists by `OR` for one particular target,
+#'    it is more intuitive and accurate to add each alternative term list to the search
+#'    term table or database directly.
+#'
 #' 2. Added `Look around` function to more accurately match SDG targets.
-#' 
-#' 
-#' 
-#' 
-#' 
-
+#'
+#'
+#'
+#'
+#'
 
 ###
 # *   --> blank
@@ -57,10 +56,10 @@ func_AND_plus <- function(v){
 
 
 ## To exclude certain terms -- used when exclude terms for a `AND` pattern by `func_AND_plus` function
-## - 1. this is an ealier version 
+## - 1. this is an ealier version
 # func_to_exclude_terms <- function(which_sdg_term, terms_to_exclude) {
 #   terms_replacement <- paste0("(?!.*", paste(terms_to_exclude, collapse = ")(?!.*"), ")")
-#   
+#
 #   if ( stringr::str_detect(which_sdg_term, "\\.\\+") )  {
 #     which_sdg_term <- gsub(pattern = "\\.\\+", replacement = terms_replacement, which_sdg_term) ## to exclude these keywords
 #   } else {
@@ -69,19 +68,19 @@ func_AND_plus <- function(v){
 #   return(which_sdg_term)
 # }
 
-## - 2. this is an updated version 
+## - 2. this is an updated version
 func_to_exclude_terms <- function(which_sdg_term, terms_to_exclude) {
   pat_exclude <- paste0("^(?!.*(?:", paste(terms_to_exclude, collapse = "|"), ")).*") ## this is more concise and efficient
   pat_and     <- which_sdg_term
-  
-  ## put `exclude` pattern before the `AND` pattern 
+
+  ## put `exclude` pattern before the `AND` pattern
   pat_exclude_and <- paste0(pat_exclude, pat_and)
   ## remove the ".*" and "^" between `exclude` pattern and the `AND` pattern
   pat_exclude_and <- gsub('\\.\\*\\^', '', pat_exclude_and)
-  
+
   return(pat_exclude_and)
 }
-### test code 
+### test code
 # exclude <- c('a', 'b')
 # terms_to_exclude <- exclude
 # func_to_exclude_terms(which_sdg_term  = SDG1_4, terms_to_exclude = exclude)
@@ -89,7 +88,76 @@ func_to_exclude_terms <- function(which_sdg_term, terms_to_exclude) {
 
 
 ## Call the `look around function`
-source('./Code/function_lookaround_nearby_n.R')
+# source('./Code/function_lookaround_nearby_n.R') ## the following code is copied from this source
[…]
