### 2023-02-18 Yingjie4Science/SDGdetector#4 @csaybar
Hi, @Yingjie4Science can you create a new release of SDGdetector to start the review? Thanks!
