+++ tests/testthat.R
@@ -1,0 +1,6 @@
+library(testthat)
+library(SDGdetector)
+
+# test_check("SDGdetector")
+
+
