+++ R/helper_SDG_search_terms.R
@@ -26,18 +26,38 @@
 # AND --> ,
 
 
-library(dplyr)
 
 
-## a vector (v) of terms to be concatenated by `AND`
-
+#' Use `AND` to Concatenate a Vector of Terms
+#'
+#' @param v a vector of characters
+#'
+#' @importFrom magrittr %>%
+#'
+#' @return A character
+#' @export
+#'
+#' @examples
+#' words <- c('apple', 'bean', 'food')
+#' func_AND_vector(v= words)
 func_AND_vector <- function(v){
   pat <- paste0("(?=.*(?:", v, "))", collapse="")
   return(pat)
 }
 
 
-## a vector (v) of terms to be concatenated by `OR`
+#' Use `OR` to Concatenate a Vector of Terms
+#'
+#' @param v a vector of characters
+#'
+#' @importFrom magrittr %>%
+#'
+#' @return A character
+#' @export
+#'
+#' @examples
+#' words <- c('apple', 'bean', 'food')
+#' func_AND_vector(v= words)
 func_OR_vector <- function(v){
   pat <- paste0(v, collapse = "|")
   pat <- paste0("(", pat, ")")
@@ -92,18 +112,26 @@ func_to_exclude_terms <- function(which_sdg_term, terms_to_exclude) {
 condition1 <- "\\s"                 ## a space after the 1st word
 condition2 <- "(?:\\w+\\s)"         ## a word character with a space after it
 
+#' @title Look Around
+#'
+#' @description Look around to match pattern in a sentence
+#'
+#' @param word_ls1 is a string, which includes a list of words connected by "|" that indicates 'OR'
+#' @param word_ls2 is a string, which includes a list of words connected by "|" that indicates 'OR'
+#' @param n        is a number, indicates the number of words to look around
+#' @param exclude  is a vector, including a list of words to be excluded from match
+#' @param third_AND_string similar to word_ls1 or word_ls2, it is a string that includes a list of words connected by "|" that indicates 'OR'
+#'
+#' @importFrom magrittr %>%
+#'
+#' @return A regex string
+#' @export
+#'
+#' @examples
+#' con1 <- c('apple', 'bean', 'food')
+#' con2 <- c('big', 'delicious')
+#' lookaround_nearby_n(word_ls1 = con1, word_ls2 = con2, n = 2, exclude = "", third_AND_string = "")
 lookaround_nearby_n <- function(word_ls1, word_ls2, n, exclude = "", third_AND_string = "") {
-
-  #'@title Look around
-  #'@description Look around to match pattern in a sentence
-  #'@author Yingjie Li
-  #'
-  #'@param word_ls1 is a string, which includes a list of words connected by "|" that indicates 'OR'
-  #'@param word_ls2 is a string, which includes a list of words connected by "|" that indicates 'OR'
-  #'@param n        is a number, indicates the number of words to look around
-  #'@param exclude  is a vector, including a list of words to be excluded from match
-  #'@param third_AND_string similar to word_ls1 or word_ls2, it is a string that includes a list of words connected by "|" that indicates 'OR'
-
   pat1 <- paste0(
     paste0('(?:', word_ls1, ')'),  ## the 1st word list
[…]
