### 2023-02-19 e7ab856
fix function `add_sdg_pattern()`
