+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(SDGdetector)
+export(add_sdg_pattern)
 export(detect_region)
 export(func_AND_vector)
 export(func_OR_vector)
