### 2023-02-19 Yingjie4Science/SDGdetector#4 @Yingjie4Science
Yes! A new release has been created. Thank you!
