+++ NEWS.md
@@ -1,0 +1,21 @@
+# SDGdetector 2.7.1
+
+* Added a `NEWS.md` file to track changes to the package.
+
+
+## New features
+
+  - added GitHub actions to automatically run R CMD check.
+    
+
+## Enhancements
+
+  - highlight novelty in the JOSS paper by comparing it to other tools, such as text2sdg.
+
+  - improved functions `sdg_color()`, `sdg_icon()`, `add_sdg_pattern()` and fixed associated documenting issues.
+  
+
+## Bug fixes
+
+  - fix GitHub installation errors.
+
