+++ .github/workflows/R-CMD-check.yaml
@@ -19,7 +19,11 @@ jobs:
       matrix:
         config:
           - {os: macOS-latest,   r: 'release'}
+          
           - {os: windows-latest, r: 'release'}
+          # Use 3.6 to trigger usage of RTools35
+          - {os: windows-latest, r: '3.6'}
+          
           - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
