### 2023-02-21 6bd4216
fix "non-ASCII characters" warnings
