+++ paper/paper.md
@@ -49,7 +49,7 @@ The global interest in moving towards a sustainable future has grown exponential
 
 The Sustainable Development Goals (SDGs) agenda, adopted by all United Nations Member States in 2015, provides a shared blueprint for nations, cities, corporations, research institutions, and individuals to track and plan their contributions to social, economic, and environmental transformations [@un_global_2019]. Although considerable efforts and contributions have been made to use existing statistical data for SDG assessments, half of the 231 indicators listed in the global indicator framework for SDGs lack either established methodologies or available data for measuring and implementing the goals ([https://unstats.un.org/sdgs/iaeg-sdgs/tier-classification](https://unstats.un.org/sdgs/iaeg-sdgs/tier-classification)). As a complement to the commonly used statistical data, textual data (e.g., websites, organization or government reports, and scientific publications) are rarely considered but show great potential for becoming a rich and important data source to narrow this existing SDG data gap [@cai_natural_2021; @chang_accelerating_2021]. For example, by identifying SDG commitments and contributions in text from legally-binding corporate annual reports, one can evaluate which SDGs are being mentioned (directly or indirectly) and to what extent corporations are moving towards them. Or, published research papers could also be evaluated to link research institutions’ commitments to SDG progress. Manually reviewing and matching text corpora to specific SDGs or targets can be extremely time-consuming and costly. In addition, though conventional manual coding may achieve high accuracy, it faces precision issues because of intercoder reliability challenges. This is especially an issue when attempting to objectively classify and map massive data into tens and hundreds of topic categories (e.g., the 169 SDG targets). To address these challenges, we developed the ***SDGdetector*** package, which automates the text analysis process via a text mining approach (Figure 1).
 
-![](figure1.png)
+![](../docs/images/figure1.png)
 
 **Figure 1.** Flowchart for identifying SDG-related statements from textual data.
 
