+++ tests/testthat/test-SDGdetector.R
@@ -1,5 +1,3 @@
-context("Detect SDGs and Targets in text")
-
 test_that("Return data frames of correct size", {
   text <- 'our goal is to mitigate climate change, end poverty, and reduce inequality globally'
   df <- data.frame(col = c(
