### 2023-02-22 bcc1dc2
fix citation and license based on CRAN comments
