+++ DESCRIPTION
@@ -22,7 +22,7 @@ Imports:
     scales,
     magick
 Depends: R (>= 3.5.0)
-License: GPL-3
+License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
