### 2023-03-01 23691a2
add a quiet/verbose option in the exported functions
