+++ R/plot_sdg_bar.R
@@ -6,6 +6,7 @@
 #' @param data  Data frame as the input
 #' @param value The value, e.g., number of SDGs, to be show in the thematic map
 #' @param sdg   Vector with SDG code to be visualized.
+#' @param quiet Logical. Suppress info message
 #'
 #' @importFrom magrittr %>%
 #' @importFrom dplyr mutate select left_join group_by
@@ -21,7 +22,7 @@
 #' @return Returns the tool text outputs.
 #' @export
 #'
-plot_sdg_bar <- function(data, sdg = "sdg", value = "value") {
+plot_sdg_bar <- function(data, sdg = "sdg", value = "value", quiet = FALSE) {
 
   k <- which(names(data) %in% c(deparse(substitute(SDG)), deparse(substitute(Value))))
   names(data)[k] <- c("sdg", "value")
