+++ R/SDGdetector.R
@@ -30,12 +30,10 @@
 #' @export
 #'
 #' @examples
-#' my_col=c('our goal is to end poverty globally', 'this product
-#' contributes to slowing down climate change')
+#' my_col <- c("our goal is to end poverty globally", "this product
+#' contributes to slowing down climate change")
 #' my_text <- data.frame(my_col)
 #' SDGdetector(my_text, my_col)
-
-
 SDGdetector <- function(x, col, quiet = FALSE) {
   nchr <- sdgs <- id <- NULL
 
@@ -64,27 +62,35 @@ SDGdetector <- function(x, col, quiet = FALSE) {
       dplyr::mutate(id = dplyr::row_number())
 
     code <- df %>%
-      dplyr::mutate(sdgs = '') ## for later use, to append data to this column
+      dplyr::mutate(sdgs = "") ## for later use, to append data to this column
 
-    for (i in 1:nrow(SDG_keys)){                           ## loop each SDG indicators
-      sdg_i_str <- SDG_keys$SDG_id[i] %>% as.character()   ## get the SDG id name
-      sdg_i_obj <- SDG_keys$SDG_keywords[i]                ## get the corresponding SDG search term list
+    for (i in 1:nrow(SDG_keys)) { ## loop each SDG indicators
+      sdg_i_str <- SDG_keys$SDG_id[i] %>% as.character() ## get the SDG id name
+      sdg_i_obj <- SDG_keys$SDG_keywords[i] ## get the corresponding SDG search term list
 
       # print(sdg_i_str)
       # print(sdg_i_obj)
 
-      code <-  code %>%
+      code <- code %>%
         as.data.frame() %>%
-
-      ## at the sentence level - count once if goals/targets are mentioned ---------------
-      dplyr::mutate(
-        match = ifelse(
-          grepl(pattern = sdg_i_obj, x = col, ignore.case = T, perl = T), 1, 0)) %>% ## yes-1 or no-0 if they match
-          # grepl(pattern = sdg_i_obj, x = as.character({{col}}), ignore.case = T, perl = T), 1, 0))  %>% ## yes-1 or no-0 if they match
+        ## at the sentence level - count once if goals/targets are mentioned -------------
         dplyr::mutate(
-          sdgs  = ifelse(match > 0, paste0(sdgs, ',', sdg_i_str), sdgs)) %>%
+          match = ifelse(
+            grepl(
+              pattern = sdg_i_obj,
+              x = col,
+              ignore.case = T,
+              perl = T
+            ),
+            1,
+            0
+          )
+        ) %>% ## yes-1 or no-0 if they match
+        dplyr::mutate(
+          sdgs = ifelse(match > 0, paste0(sdgs, ",", sdg_i_str), sdgs)
+        ) %>%
         dplyr::select(-match) %>% ## remove this column
-        dplyr::mutate(sdgs = gsub("^,*|(?<=,),|,*$", "", sdgs, perl=T)) %>%
+        dplyr::mutate(sdgs = gsub("^,*|(?<=,),|,*$", "", sdgs, perl = T)) %>%
         as.data.frame()
     }
 
@@ -93,43 +99,54 @@ SDGdetector <- function(x, col, quiet = FALSE) {
   } else {
     ##
     df <- x %>%
-      dplyr::rename(col = {{col}}) %>%
+      dplyr::rename(col = {{ col }}) %>%
       dplyr::mutate(id = dplyr::row_number())
 
     ## check the number of characters in the sentence
     df_nchar <- df %>%
       dplyr::mutate(nchr = stringr::str_length(col)) %>%
       dplyr::filter(nchr > 750)
-    if(nrow(df_nchar) > 0){
-      message(paste0("The length of your input text reached the limit in PCRE, ",
[…]

[…]
