+++ man/sdg_icon.Rd
@@ -9,7 +9,8 @@ sdg_icon(x, res = 200)
 \arguments{
 \item{x}{Numeric code for each SDG, ranging from 1 to 17}
 
-\item{res}{Resolution of SDG icon. Default: \code{res = 200} indicates resizing proportionally to 200px}
+\item{res}{Resolution of SDG icon. Default: \code{res = 200} indicates
+resizing proportionally to 200px}
 }
 \description{
 The \code{sdg_icon}  function provides the specific icon for each SDG
