### 2023-03-06 Yingjie4Science/SDGdetector#4 @csaybar
Hi @Yingjie4Science,

Thank for you your reply!
You can check the current code coverage running:

"""r
library(covr)
covr::package_coverage()

SDGdetector Coverage: 18.78%
R/add_sdg_pattern.R: 0.00%
R/detect_region.R: 0.00%
R/helper_SDG_search_terms.R: 0.00%
R/plot_sdg_bar.R: 0.00%
R/plot_sdg_map.R: 0.00%
R/sdg_color.R: 0.00%
R/sdg_icon.R: 0.00%
R/SDGdetector.R: 84.31%
"""

https://covr.r-lib.org/
