+++ docs/accuracy_evaluation/validation_analysis.Rmd
@@ -1,0 +1,323 @@
+---
+title: "Text_mining"
+author: "Yingjie"
+date: "01/23/2022"
+output: html_document
+editor_options: 
+  chunk_output_type: inline
+---
+
+
+  This script aims to use human intelligence to evaluate the accuracy of the text mining results. 
+  For details, please refer to [this report](https://drive.google.com/file/d/1EHUV6Jc3N4A-IshKU4dbxtIqlfj50mzi/view)
+  
+  
+# Directories and packages
+
+```{r include=FALSE}
+### To clear your environment
+remove(list = ls())
+
+
+# directories
+dir.this <- "./docs/accuracy_evaluation/"
+dir.data <- paste0(dir.this, "data/")
+dir.figures <- dir.this
+
+# load packages
+library(dplyr)
+library(readxl)
+library(writexl)
+
+library(tidyverse) # piping and wrangling
+library(ggplot2)
+library(ggpubr)
+```
+
+
+
+# Data
+
+  We validated the accuracy by conducting three round of test. For each round, we asked 
+  four experts to identify if the machine-coded results are correct or not. 
+  
+  The data are save in XLSX files under each round's folder. 
+  
+  ```
+  
+  ./docs/accuracy_evaluation/
+      |
+      |__ data/
+      |    |
+      |    |__ task1_R1_done/
+      |    |
+      |    |__ task1_R2_done/
+      |    |
+      |    |__ task1_R3_done/
+      |
+      |
+      |__ (results and figures)
+      
+  ```
+  
+## 1. Clean data from each round
+
+  First, we need to 
+    - read in and clean four coders' data
+    - calculate the accuracy based on each coder's evaluation 
+    - check alignment across coders 
+    - save result in xlsx for visualization 
+  
+  Please modify `nth_round` and run all the code in this section 
+  (i.e., Section ## 1. Clean data from each round) before running code in the 2nd section
+  for visualization. In this case, you will need to run code in section 1 for three times. 
+  
+```{r - read and clean}
+getwd()
+
+
+## choose the round number and run all code in section 1.
+nth_round <- 1
+# nth_round <- 2
+# nth_round <- 3
+
+
+## list data from this round of evaluation -----------------------------------------------
+dir.task1 <- paste0(dir.data, "task1_R", nth_round, "_done")
+dir.task1
+f.ls <- list.files(path = dir.task1, pattern = "task1_", full.names = T, recursive = T)
+f.ls
+## test line
+# f    <- "./Data/data_TextMining/coded_validation_coder/planB_done/task1/task1_YL.xlsx"
+
+
+## read in all the XLSX ------------------------------------------------------------------
+ds <- data.frame()
+for (f in f.ls) {
+  coder <- basename(f) %>% gsub("task1_|\\.xlsx", "", .)
+  coder
+  d <- readxl::read_excel(path = f) %>%
+    dplyr::mutate(coder = coder, index = as.numeric(row.names(.)))
+  ds <- rbind(ds, d)
+}
+
+unique(ds$correct_or_not)
+unique(ds$coder)
+
+
+## unify the checking result to "0, 1, NA" only ------------------------------------------
+ds0 <- ds %>%
[…]
