### 2023-03-06 Yingjie4Science/SDGdetector#4 @Yingjie4Science
Thanks, @csaybar ! This is super helpful. We have added more tests, and the current coverage is
`> covr::package_coverage()
SDGdetector Coverage: 93.45%
R/SDGdetector.R: 82.05%
R/add_sdg_pattern.R: 95.45%
R/helper_SDG_search_terms.R: 95.45%
R/plot_sdg_bar.R: 97.30%
R/detect_region.R: 100.00%
R/plot_sdg_map.R: 100.00%
R/sdg_color.R: 100.00%
R/sdg_icon.R: 100.00%`
