### 2023-03-06 7ea64b3
add a warning message to `plot_sdg_map()`
