### 2023-03-06 bfed791
add automated tests for all functions, now test coverage=93.45%
