### 2023-03-16 Yingjie4Science/SDGdetector#4 @csaybar
Hi, @Yingjie4Science, I have finished up the checklist at openjournals/joss-reviews#5124 and as far as my review is concerned, you are all set. Your edits look very good in both the paper/code.
