### 2023-03-16 #5124 @csaybar
Hello @fboehm, I wanted to let you know that I have completed my review of https://github.com/Yingjie4Science/SDGdetector/issues/4. I appreciate the opportunity you have given me to review this paper!
