### 2023-03-17 Yingjie4Science/SDGdetector#4 @Yingjie4Science
Hi @csaybar , thank you again for your time and very constructive comments! I learned many new things from you and the process!
