### 2023-03-25 d2cc061
update author and CITATION info
