### 2023-03-27 1ada0ca
proof and minor edits on affiliations and text
