+++ NEWS.md
@@ -1,21 +1,12 @@
 # SDGdetector 2.7.1
 
-* Added a `NEWS.md` file to track changes to the package.
+* the version for JOSS publication.
 
 
-## New features
+## New edits
 
-  - added GitHub actions to automatically run R CMD check.
-    
-
-## Enhancements
-
-  - highlight novelty in the JOSS paper by comparing it to other tools, such as text2sdg.
-
-  - improved functions `sdg_color()`, `sdg_icon()`, `add_sdg_pattern()` and fixed associated documenting issues.
-  
-
-## Bug fixes
-
-  - fix GitHub installation errors.
+  - fix author affiliation information
+  - fix typos
+  - add automated tests for all functions, now test coverage=93.45%, per reviewers' comments
+  - added instructions in the paper for replicating the accuracy assessment, per reviewers' comments
 
